requirements.txt
# requirements.txt
google-generativeai>=0.3.0
python-dotenv>=1.0.0
uvicorn>=0.24.0
fastapi>=0.104.0
requests>=2.31.0
aiohttp>=3.9.0
pydantic>=2.0.0
reportlab>=4.0.0
